requirements-dev.txt
# Development Dependencies for Divyadrishti
# Install with: pip install -r requirements-dev.txt

# Testing Framework
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
pytest-flask>=1.2.0
pytest-asyncio>=0.21.1

# Code Quality & Linting
black>=23.7.0
isort>=5.12.0
flake8>=6.0.0
flake8-docstrings>=1.7.0
flake8-import-order>=0.18.2
flake8-bugbear>=23.7.10

# Type Checking
mypy>=1.5.0
types-requests>=2.31.0
types-Flask>=1.1.6

# Security Scanning
bandit>=1.7.5
safety>=2.3.5

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0
sphinx-autodoc-typehints>=1.24.0
mkdocs>=1.5.0
mkdocs-material>=9.1.21

# Pre-commit Hooks
pre-commit>=3.3.3

# Development Tools
ipython>=8.14.0
ipdb>=0.13.13
python-dotenv>=1.0.0

# Performance Profiling
py-spy>=0.3.14
memory-profiler>=0.61.0
line-profiler>=4.1.1

# Database Tools
alembic>=1.11.1
sqlalchemy-utils>=0.41.1

# API Testing
httpx>=0.24.1
responses>=0.23.1
requests-mock>=1.11.0

# Load Testing
locust>=2.15.1

# Monitoring & Logging
prometheus-client>=0.17.1
structlog>=23.1.0

# Development Server
watchdog>=3.0.0
gunicorn>=21.2.0
flask-debugtoolbar>=0.13.1

# Build Tools
build>=0.10.0
twine>=4.0.2

# Jupyter Notebooks (for data analysis)
jupyter>=1.0.0
notebook>=7.0.0
jupyterlab>=4.0.0

# Data Analysis (for analytics features)
pandas>=2.0.3
numpy>=1.24.3
matplotlib>=3.7.2
seaborn>=0.12.2

# GIS Development Tools
geopandas>=0.13.2
shapely>=2.0.1
fiona>=1.9.4
pyproj>=3.6.0
rasterio>=1.3.8

# Environment Management
python-decouple>=3.8
environs>=9.5.0

# Debugging
pdb++>=0.10.3
better-exceptions>=0.3.3
icecream>=2.1.3

# Code Coverage Reporting
coverage[toml]>=7.2.7
codecov>=2.1.13

# Configuration Management
dynaconf>=3.2.0

# Mock Services
responses>=0.23.1
requests-mock>=1.11.0

# Performance Testing
pytest-benchmark>=4.0.0

# Selenium for E2E Testing
selenium>=4.11.0
webdriver-manager>=3.8.6

# API Documentation
flasgger>=0.9.7.1
apispec>=6.3.0

# Development Utilities
rich>=13.4.2
click>=8.1.6
tqdm>=4.65.0

# File Watching
watchfiles>=0.19.0

# Code Complexity Analysis
radon>=6.0.1
xenon>=0.9.0

# Dependency Management
pip-tools>=7.1.0
pipdeptree>=2.9.6

# License Checking
pip-licenses>=4.3.0

# Vulnerability Scanning
pip-audit>=2.6.1

# Development Containers
docker-compose>=1.29.2

# Browser Automation
playwright>=1.36.0

# Performance Monitoring
psutil>=5.9.5

# Development Helpers
python-json-logger>=2.0.7
colorlog>=6.7.0

# Testing Utilities
factory-boy>=3.3.0
faker>=19.3.0

# Development Configuration
python-box>=7.0.1

# Development Validation
schema>=0.7.5
pydantic>=2.1.1

# Development Caching
diskcache>=5.6.1
cachetools>=5.3.1

# Development Networking
httpx>=0.24.1
aiohttp>=3.8.5

# Development Async
aiofiles>=23.1.0

# Development Monitoring
sentry-sdk>=1.29.2

# Development Logging
loguru>=0.7.0

# Development CLI
typer>=0.9.0
rich-click>=1.6.1

# Development Templates
jinja2>=3.1.2

# Development Utilities
more-itertools>=10.1.0
python-slugify>=8.0.1
python-dateutil>=2.8.2

# Development Cryptography
cryptography>=41.0.3
passlib>=1.7.4

# Development Authentication
pyjwt>=2.8.0

# Development Parsing
beautifulsoup4>=4.12.2
lxml>=4.9.3

# Development Text Processing
textdistance>=4.6.0

# Development File Processing
openpyxl>=3.1.2
xlsxwriter>=3.1.2

# Development PDF Processing
pypdf2>=3.0.1
reportlab>=4.0.4

# Development Image Processing
pillow>=10.0.0
imageio>=2.31.1

# Development Encoding
chardet>=5.1.0
charset-normalizer>=3.2.0

# Development WebSockets
websockets>=11.0.3

# Development API
fastapi>=0.101.0

# Development Database
redis>=4.6.0

# Development Code Metrics
vulture>=2.9.1

# Development Workflow
invoke>=2.2.0

# Development Time Series
arrow>=1.2.3

# Development Serialization
orjson>=3.9.4

# Development Hashing
xxhash>=3.2.0

# Development Compression
zstandard>=0.21.0

# Development DNS
dnspython>=2.4.0

# Development Network
netaddr>=0.8.0

# Development Configuration Validation
cerberus>=1.3.4

# Development Code Generation
cookiecutter>=2.2.3

# Development Git Tools
gitpython>=3.1.32

# Development Memory Debugging
pympler>=0.9

# Development Import Sorting
isort[colors]>=5.12.0
